feat: refactor into modules
