+++ src/modules/verification/verifyUser.ts
@@ -1,7 +1,7 @@
 import { Guild, GuildMember } from "discord.js";
 
-import { logHandler } from "../utils/logHandler";
-import { sendLogMessage } from "../utils/sendLogMessage";
+import { logHandler } from "../../utils/logHandler";
+import { sendLogMessage } from "../../utils/sendLogMessage";
 
 /**
  * Grants the verified role to a user.
